Fixed upload zip, prefer PNG.

+++ src/com/esotericsoftware/clippy/Upload.java
@@ -17,7 +17,6 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLEncoder;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -25,7 +24,8 @@ import javax.imageio.IIOImage;
 import javax.imageio.ImageIO;
 import javax.imageio.ImageWriteParam;
 import javax.imageio.ImageWriter;
-import javax.imageio.stream.ImageOutputStream;
+import javax.imageio.plugins.jpeg.JPEGImageWriteParam;
+import javax.imageio.stream.FileImageOutputStream;
 
 import com.esotericsoftware.clippy.util.Imgur.ImageResponse;
 import com.esotericsoftware.clippy.util.Imgur.ImgurAPI;
@@ -258,7 +258,7 @@ public abstract class Upload {
 		}
 	}
 
-	static public void uploadImage (BufferedImage image) {
+	static public void uploadImage (BufferedImage image, boolean forcePNG) {
 		if (clippy.imageUpload == null) return;
 		int number = Util.nextUploadID();
 		File filePNG = null;
@@ -273,24 +273,26 @@ public abstract class Upload {
 		}
 
 		File fileJPG = null;
-		try {
-			fileJPG = Util.nextUploadFile(number, ".jpg");
-			if (TRACE) trace("Writing JPG file: " + fileJPG);
-			Iterator<ImageWriter> writers = ImageIO.getImageWritersByFormatName("jpg");
-			ImageWriter writer = writers.next();
-			ImageWriteParam param = writer.getDefaultWriteParam();
-			param.setCompressionMode(ImageWriteParam.MODE_EXPLICIT);
-			param.setCompressionQuality(0.87f);
-			ImageOutputStream output = ImageIO.createImageOutputStream(fileJPG);
-			if (output == null) throw new RuntimeException("Unknown image format: " + fileJPG);
-			writer.setOutput(output);
-			writer.write(null, new IIOImage(image, null, null), param);
-			writer.dispose();
-			output.close();
-		} catch (IOException ex) {
-			if (ERROR) error("Error writing JPG file: " + fileJPG, ex);
-			if (fileJPG != null) fileJPG.delete();
-			fileJPG = null;
+		if (!forcePNG) {
+			try {
+				fileJPG = Util.nextUploadFile(number, ".jpg");
+				if (TRACE) trace("Writing JPG file: " + fileJPG);
+
+				JPEGImageWriteParam param = new JPEGImageWriteParam(null);
+				param.setCompressionMode(ImageWriteParam.MODE_EXPLICIT);
+				param.setCompressionQuality(1);
+
+				FileImageOutputStream output = new FileImageOutputStream(fileJPG);
+				ImageWriter writer = ImageIO.getImageWritersByFormatName("jpg").next();
+				writer.setOutput(output);
+				writer.write(null, new IIOImage(image, null, null), param);
+				writer.dispose();
+				output.close();
+			} catch (IOException ex) {
+				if (ERROR) error("Error writing JPG file: " + fileJPG, ex);
+				if (fileJPG != null) fileJPG.delete();
+				fileJPG = null;
+			}
 		}
 
 		File file;
@@ -299,12 +301,15 @@ public abstract class Upload {
 			file = fileJPG;
 		} else if (fileJPG == null)
 			file = filePNG;
-		else if (fileJPG.length() < filePNG.length() * 0.9d) {
-			file = fileJPG;
-			filePNG.delete();
-		} else {
-			file = filePNG;
-			fileJPG.delete();
+		else {
+			long lengthPNG = filePNG.length();
+			if (lengthPNG > 350000 && fileJPG.length() / (double)lengthPNG < 0.66) {
+				file = fileJPG;
+				filePNG.delete();
+			} else {
+				file = filePNG;
+				fileJPG.delete();
+			}
 		}
 
 		clippy.imageUpload.upload(file, true, new UploadListener() {
@@ -334,7 +339,6 @@ public abstract class Upload {
 				Paths paths = new Paths();
 				for (String path : files) {
 					File file = new File(path);
-					paths.addFile(path);
 					paths.glob(file.getParent(), file.getName() + "/**");
 				}
 				Scar.zip(paths, zip.getAbsolutePath());
